+++ volmdlr/faces.py
@@ -8,7 +8,7 @@ import warnings
 from itertools import chain, product
 from typing import List
 import matplotlib.pyplot as plt
-import numpy as npy
+import numpy as np
 
 from dessia_common.core import DessiaObject
 
@@ -21,7 +21,7 @@ import volmdlr.curves as volmdlr_curves
 import volmdlr.geometry
 import volmdlr.grid
 from volmdlr import surfaces
-from volmdlr.utils.parametric import array_range_search
+from volmdlr.utils.parametric import update_face_grid_points_with_inner_polygons
 import volmdlr.wires
 
 warnings.simplefilter("once")
@@ -94,9 +94,7 @@ class Face3D(volmdlr.core.Primitive3D):
         if not self.bounding_box.point_belongs(point3d, 1e-3):
             return False
         point2d = self.surface3d.point3d_to_2d(point3d)
-        # check_point3d = self.surface3d.point2d_to_3d(point2d)
-        # if check_point3d.point_distance(point3d) > tol:
-        if not self.surface3d.point_on_surface(point3d, tol):
+        if not self.surface3d.point_belongs(point3d, tol):
             return False
 
         return self.surface2d.point_belongs(point2d)
@@ -337,6 +335,55 @@ class Face3D(volmdlr.core.Primitive3D):
         Specifies the number of subdivision when using triangulation by lines. (Old triangulation).
         """
         return [], []
+
+    def grid_points(self, grid_size, polygon_data=None):
+        """
+        Parametric tesselation points.
+        """
+        if polygon_data:
+            outer_polygon, inner_polygons = polygon_data
+        else:
+            outer_polygon, inner_polygons = self.get_face_polygons()
+
+        u, v = self._get_grid_axis(outer_polygon, grid_size)
+        if not u or not v:
+            return []
+        if inner_polygons:
+            grid_points = []
+            points_indexes_map = {}
+            for j, v_j in enumerate(v):
+                for i, u_i in enumerate(u):
+                    grid_points.append((u_i, v_j))
+                    points_indexes_map[(i, j)] = len(grid_points) - 1
+            grid_points = update_face_grid_points_with_inner_polygons(inner_polygons,
+                                                                      [grid_points, u, v, points_indexes_map])
+        else:
+            grid_points = np.array([[u_i, v_j] for v_j in v for u_i in u], dtype=np.float64)
+        grid_points = self._update_grid_points_with_outer_polygon(outer_polygon, grid_points)
+
+        return grid_points
+
+    @staticmethod
+    def _get_grid_axis(outer_polygon, grid_size):
+        """Helper function to grid_points."""
+        u_min, u_max, v_min, v_max = outer_polygon.bounding_rectangle.bounds()
+        number_points_u, number_points_v = grid_size
+        u_step = (u_max - u_min) / (number_points_u - 1) if number_points_u > 1 else (u_max - u_min)
+        v_step = (v_max - v_min) / (number_points_v - 1) if number_points_v > 1 else (v_max - v_min)
+        u = [u_min + i * u_step for i in range(number_points_u)]
+        v = [v_min + i * v_step for i in range(number_points_v)]
+
+        return u, v
+
+    @staticmethod
+    def _update_grid_points_with_outer_polygon(outer_polygon, grid_points):
+        """Helper function to grid_points."""
+        # Find the indices where points_in_polygon is True (i.e., points inside the polygon)
+        indices = np.where(outer_polygon.points_in_polygon(grid_points, include_edge_points=False) == 0)[0]
+        grid_points = np.delete(grid_points, indices, axis=0)
+        polygon_points = set(outer_polygon.points)
+        points = [volmdlr.Point2D(*point) for point in grid_points if volmdlr.Point2D(*point) not in polygon_points]
+        return points
 
     def get_face_polygons(self):
         """Get face polygons."""
@@ -1241,24 +1288,24 @@ class Face3D(volmdlr.core.Primitive3D):
             {point for face in faces1 for point in face.points} for _, faces1 in face_decomposition1.items()
         ]
         list_set_points1 = [
-            npy.array([(point[0], point[1], point[2]) for point in sets_points1]) for sets_points1 in list_set_points1
+            np.array([(point[0], point[1], point[2]) for point in sets_points1]) for sets_points1 in list_set_points1
         ]
         list_set_points2 = [
             {point for face in faces2 for point in face.points} for _, faces2 in face_decomposition2.items()
         ]
         list_set_points2 = [
-            npy.array([(point[0], point[1], point[2]) for point in sets_points2]) for sets_points2 in list_set_points2
+            np.array([(point[0], point[1], point[2]) for point in sets_points2]) for sets_points2 in list_set_points2
         ]
 
         minimum_distance = math.inf
         index1, index2 = None, None
         for sets_points1, sets_points2 in product(list_set_points1, list_set_points2):
-            distances = npy.linalg.norm(sets_points2[:, npy.newaxis] - sets_points1, axis=2)
-            sets_min_dist = npy.min(distances)
+            distances = np.linalg.norm(sets_points2[:, np.newaxis] - sets_points1, axis=2)
+            sets_min_dist = np.min(distances)
             if sets_min_dist < minimum_distance:
                 minimum_distance = sets_min_dist
-                index1 = next((i for i, x in enumerate(list_set_points1) if npy.array_equal(x, sets_points1)), -1)
-                index2 = next((i for i, x in enumerate(list_set_points2) if npy.array_equal(x, sets_points2)), -1)
+                index1 = next((i for i, x in enumerate(list_set_points1) if np.array_equal(x, sets_points1)), -1)
+                index2 = next((i for i, x in enumerate(list_set_points2) if np.array_equal(x, sets_points2)), -1)
         faces1 = list(face_decomposition1.values())[index1]
         faces2 = list(face_decomposition2.values())[index2]
 
@@ -1332,18 +1379,18 @@ class Face3D(volmdlr.core.Primitive3D):
             {point for face in faces1 for point in face.points} for _, faces1 in face_decomposition1.items()
         ]
         list_set_points1 = [
-            npy.array([(point[0], point[1], point[2]) for point in sets_points1]) for sets_points1 in list_set_points1
+            np.array([(point[0], point[1], point[2]) for point in sets_points1]) for sets_points1 in list_set_points1
         ]
-        list_set_points2 = [npy.array([(point[0], point[0], point[0])])]
+        list_set_points2 = [np.array([(point[0], point[0], point[0])])]
 
         minimum_distance = math.inf
         index1 = None
         for sets_points1, sets_points2 in product(list_set_points1, list_set_points2):
-            distances = npy.linalg.norm(sets_points2[:, npy.newaxis] - sets_points1, axis=2)
-            sets_min_dist = npy.min(distances)
+            distances = np.linalg.norm(sets_points2[:, np.newaxis] - sets_points1, axis=2)
+            sets_min_dist = np.min(distances)
             if sets_min_dist < minimum_distance:
                 minimum_distance = sets_min_dist
-                index1 = next((i for i, x in enumerate(list_set_points1) if npy.array_equal(x, sets_points1)), -1)
+                index1 = next((i for i, x in enumerate(list_set_points1) if np.array_equal(x, sets_points1)), -1)
         return list(face_decomposition1.values())[index1]
 
     def point_distance(self, point, return_other_point: bool = False):
@@ -1946,6 +1993,42 @@ class PlaneFace3D(Face3D):
         return cls(plane3d, surface, name)
 
 
+class PeriodicalFaceMixin:
+    """
+    Abstract class for mutualizing methods for faces constructed on periodic surfaces.
+    """
+
+    def point_belongs(self, point3d: volmdlr.Point3D, tol: float = 1e-6) -> bool:
+        """
+        Checks if a 3D point lies on the face.
+
+        :param point3d: The 3D point to be checked.
+        :type point3d: volmdlr.Point3D
+
+        :param tol: Tolerance for the check.
+        :type tol: float, optional
+
+        :return: True if the point is on the ConicalFace3D, False otherwise.
+        :rtype: bool
+        """
+        if not self.surface3d.point_belongs(point3d, tol):
+            return False
+        point2d = self.surface3d.point3d_to_2d(point3d)
+        u_min, u_max, v_min, v_max = self.surface2d.bounding_rectangle().bounds()
+        if self.surface3d.x_periodicity:
+            if point2d.x < u_min - tol:
+                point2d.x += self.surface3d.x_periodicity
+            elif point2d.x > u_max + tol:
+                point2d.x -= self.surface3d.x_periodicity
+        if self.surface3d.y_periodicity:
+            if point2d.y < v_min - tol:
+                point2d.y += self.surface3d.y_periodicity
+            elif point2d.y > v_max + tol:
+                point2d.y -= self.surface3d.y_periodicity
+
+        return self.surface2d.point_belongs(point2d)
+
+
 class Triangle3D(PlaneFace3D):
     """
     Defines a Triangle3D class.
@@ -2282,7 +2365,7 @@ class Triangle3D(PlaneFace3D):
         return self.planeface_minimum_distance(triangle_face, return_points)
 
 
-class CylindricalFace3D(Face3D):
+class CylindricalFace3D(PeriodicalFaceMixin, Face3D):
     """
     Defines a CylindricalFace3D class.
 
@@ -2340,22 +2423,7 @@ class CylindricalFace3D(Face3D):
             lines.append(volmdlr_curves.Line2D(volmdlr.Point2D(theta, zmin), volmdlr.Point2D(theta, zmax)))
         return lines, []
 
-    def point_belongs(self, point3d: volmdlr.Point3D, tol: float = 1e-6):
-        """
-        Tells you if a point is on the 3D Cylindrical face and inside its contour.
-        """
-        point2d = self.surface3d.point3d_to_2d(point3d)
-        point2d_plus_2pi = point2d.translation(volmdlr.Point2D(volmdlr.TWO_PI, 0))
-        point2d_minus_2pi = point2d.translation(volmdlr.Point2D(-volmdlr.TWO_PI, 0))
-        # check_point3d = self.surface3d.point2d_to_3d(point2d)
-        # if check_point3d.point_distance(point3d) > tol:
-        if not self.surface3d.point_on_surface(point3d, tol):
-            return False
-        return any(self.surface2d.point_belongs(pt2d) for pt2d in [point2d, point2d_plus_2pi, point2d_minus_2pi])
-
     def parametrized_grid_size(self, angle_resolution, z_resolution):
-        # angle_resolution = 5
-        # z_resolution = 0
         theta_min, theta_max, zmin, zmax = self.surface2d.bounding_rectangle().bounds()
         delta_theta = theta_max - theta_min
         number_points_x = max(angle_resolution, int(delta_theta * angle_resolution))
@@ -2425,6 +2493,11 @@ class CylindricalFace3D(Face3D):
         return True
 
     def planeface_intersections(self, planeface: PlaneFace3D):
+        """
+        Finds intersections with the given plane face.
+
+        :param planeface: Plane face to evaluate the intersections.
+        """
         planeface_intersections = planeface.cylindricalface_intersections(self)
         return planeface_intersections
 
@@ -2459,7 +2532,7 @@ class CylindricalFace3D(Face3D):
         return volmdlr.wires.Wire3D([vme.LineSegment3D(point1, point2)])
 
 
-class ToroidalFace3D(Face3D):
+class ToroidalFace3D(PeriodicalFaceMixin, Face3D):
     """
     Defines a ToroidalFace3D class.
 
@@ -2630,7 +2703,7 @@ class ToroidalFace3D(Face3D):
         return planeface_intersections
 
 
-class ConicalFace3D(Face3D):
+class ConicalFace3D(PeriodicalFaceMixin, Face3D):
     """
     Defines a ConicalFace3D class.
 
@@ -2642,13 +2715,7 @@ class ConicalFace3D(Face3D):
 
     """
 
-    min_x_density = 5
-    min_y_density = 1
-
     def __init__(self, surface3d: surfaces.ConicalSurface3D, surface2d: surfaces.Surface2D, name: str = ""):
-        surface2d_br = surface2d.bounding_rectangle()
-        if surface2d_br[0] < 0:
-            surface2d = surface2d.translation(volmdlr.Vector2D(2 * math.pi, 0))
         Face3D.__init__(self, surface3d=surface3d, surface2d=surface2d, name=name)
         self._bbox = None
 
@@ -2709,8 +2776,6 @@ class ConicalFace3D(Face3D):
         Cut a rectangular piece of the ConicalSurface3D object and return a ConicalFace3D object.
 
         """
-        # theta1 = angle_principal_measure(theta1)
-        # theta2 = angle_principal_measure(theta2)
         if theta1 < 0:
             theta1, theta2 = theta1 + 2 * math.pi, theta2 + 2 * math.pi
         if theta1 == theta2:
@@ -2761,28 +2826,6 @@ class ConicalFace3D(Face3D):
         point2 = self.surface3d.frame.origin + self.surface3d.frame.w * zmax
         return volmdlr.wires.Wire3D([vme.LineSegment3D(point1, point2)])
 
-    def point_belongs(self, point3d: volmdlr.Point3D, tol: float = 1e-6):
-        """
-        Tells you if a point is on the 3D conical face and inside its contour.
-        """
-        if not self.bounding_box.point_belongs(point3d):
-            return False
-        x, y, z = self.surface3d.frame.global_to_local_coordinates(point3d)
-        radius = z * math.tan(self.surface3d.semi_angle)
-        point2d = volmdlr.Point2D(0, z)
-        if radius != 0.0:
-            theta = volmdlr.geometry.sin_cos_angle(x / radius, y / radius)
-            if abs(theta) < 1e-9:
-                theta = 0.0
-            point2d = volmdlr.Point2D(theta, z)
-
-        point2d_plus_2pi = point2d.translation(volmdlr.Point2D(volmdlr.TWO_PI, 0))
-        check_point3d = self.surface3d.point2d_to_3d(self.surface3d.point3d_to_2d(point3d))
-        if check_point3d.point_distance(point3d) > tol:
-            return False
-
-        return self.surface2d.point_belongs(point2d) or self.surface2d.point_belongs(point2d_plus_2pi)
-
     def circle_inside(self, circle: volmdlr_curves.Circle3D):
         """
         Verifies if a circle 3D lies completely on the Conical face.
@@ -2799,7 +2842,7 @@ class ConicalFace3D(Face3D):
         return True
 
 
-class SphericalFace3D(Face3D):
+class SphericalFace3D(PeriodicalFaceMixin, Face3D):
     """
     Defines a SpehericalFace3D class.
 
@@ -2862,6 +2905,50 @@ class SphericalFace3D(Face3D):
         number_points_y = int(delta_phi * angle_resolution)
 
         return number_points_x, number_points_y
+
+    def grid_points(self, grid_size, polygon_data=None):
+        """
+        Parametric tesselation points.
+        """
+        if polygon_data:
+            outer_polygon, inner_polygons = polygon_data
+        else:
+            outer_polygon, inner_polygons = self.get_face_polygons()
+        u, v, u_size, _ = self._get_grid_axis(outer_polygon, grid_size)
+        if not u or not v:
+            return []
+        if inner_polygons:
+            points = []
+            points_indexes_map = {}
+            for j, v_j in enumerate(v):
+                for i in range(u_size):
+                    if (j % 2 == 0 and i % 2 == 0) or (j % 2 != 0 and i % 2 != 0):
+                        points.append((u[i], v_j))
+                        points_indexes_map[(i, j)] = len(points) - 1
+            points = np.array(points, dtype=np.float64)
+            points = update_face_grid_points_with_inner_polygons(inner_polygons, [points, u, v, points_indexes_map])
+        else:
+            points = np.array([(u[i], v_j) for j, v_j in enumerate(v) for i in range(u_size) if
+                                   (j % 2 == 0 and i % 2 == 0) or (j % 2 != 0 and i % 2 != 0)],
+                                   dtype=np.float64)
+
+        points = self._update_grid_points_with_outer_polygon(outer_polygon, points)
+
+        return points
+
+    @staticmethod
+    def _get_grid_axis(outer_polygon, grid_size):
+        """Helper function to grid_points."""
+        theta_min, theta_max, phi_min, phi_max = outer_polygon.bounding_rectangle.bounds()
+        theta_resolution, phi_resolution = grid_size
+        step_u = 0.5 * math.radians(theta_resolution)
+        step_v = math.radians(phi_resolution)
+        u_size = math.ceil((theta_max - theta_min) / step_u)
+        v_size = math.ceil((phi_max - phi_min) / step_v)
+        v_start = phi_min + step_v
+        u = [theta_min + step_u * i for i in range(u_size)]
+        v = [v_start + j * step_v for j in range(v_size - 1)]
+        return u, v, u_size, v_size
 
     @classmethod
     def from_surface_rectangular_cut(
@@ -3194,12 +3281,7 @@ class BSplineFace3D(Face3D):
                 number_points_x, number_points_y = 5, 3
             else:
                 number_points_x, number_points_y = 3, 5
-            outer_polygon = self.surface2d.outer_contour.to_polygon(angle_resolution=15, discretize_line=True)
-            points_grid, x, y, grid_point_index = outer_polygon.grid_triangulation_points(
-                number_points_x, number_points_y, include_edge_points=False
-            )
-            if self.surface2d.inner_contours:
-                points_grid = self._get_bbox_inner_contours_points(points_grid, x, y, grid_point_index)
+            points_grid = self.grid_points([number_points_x, number_points_y])
             points3d = [self.surface3d.point2d_to_3d(point) for point in points_grid]
         except ZeroDivisionError:
             points3d = []
@@ -3210,22 +3292,6 @@ class BSplineFace3D(Face3D):
         return volmdlr.core.BoundingBox.from_bounding_boxes(
             [volmdlr.core.BoundingBox.from_points(points3d), self.outer_contour3d.bounding_box]
         ).scale(1.001)
-
-    def _get_bbox_inner_contours_points(self, points_grid, x, y, grid_point_index):
-        """Helper function to get_bounding_box."""
-        for inner_contour in self.surface2d.inner_contours:
-            inner_polygon = inner_contour.to_polygon(angle_resolution=5, discretize_line=True)
-            # removes with a region search the grid points that are in the inner contour
-            xmin, xmax, ymin, ymax = inner_polygon.bounding_rectangle.bounds()
-            for i in array_range_search(x, xmin, xmax):
-                for j in array_range_search(y, ymin, ymax):
-                    point = grid_point_index.get((i, j))
-                    if not point:
-                        continue
-                    if inner_polygon.point_belongs(point):
-                        points_grid.remove(point)
-                        grid_point_index.pop((i, j))
-        return points_grid
 
     def triangulation_lines(self, resolution=25):
         """
@@ -3378,7 +3444,9 @@ class BSplineFace3D(Face3D):
         return corresponding_directions, grid2d_direction
 
     def adjacent_direction_uv(self, other_bspline_face3d, corresponding_directions):
-
+        """
+        Returns the sides that are adjacents to other BSpline face.
+        """
         extremities = self.extremities(other_bspline_face3d)
         start1, start2 = extremities[0], extremities[2]
         borders_points = [volmdlr.Point2D(0, 0), volmdlr.Point2D(1, 0), volmdlr.Point2D(1, 1), volmdlr.Point2D(0, 1)]
@@ -3470,7 +3538,7 @@ class BSplineFace3D(Face3D):
         shared = []
         for k, point1 in enumerate(contour1.primitives):
             if dis_sorted[0] == dis_sorted[1]:
-                indices = npy.where(npy.array(dis) == dis_sorted[0])[0]
+                indices = np.where(np.array(dis) == dis_sorted[0])[0]
                 index1 = indices[0]
                 index2 = indices[1]
             else:
@@ -3676,8 +3744,8 @@ class BSplineFace3D(Face3D):
             return v_centers
         if u_radius and not v_radius:
             return u_centers
-        u_mean = npy.mean(u_radius)
-        v_mean = npy.mean(v_radius)
+        u_mean = np.mean(u_radius)
+        v_mean = np.mean(v_radius)
         if u_mean > v_mean:
             return v_centers
         if u_mean < v_mean: